+++ packages/admin/cms-admin/src/generator/future/generateForm/generateFormField.ts
@@ -1,11 +1,62 @@
-import { IntrospectionEnumType, IntrospectionNamedTypeRef, IntrospectionObjectType, IntrospectionQuery } from "graphql";
+import { IntrospectionEnumType, IntrospectionInputValue, IntrospectionNamedTypeRef, IntrospectionObjectType, IntrospectionQuery } from "graphql";
 
-import { FormConfig, FormFieldConfig } from "../generator";
+import { Prop } from "../generateForm";
+import { FormConfig, FormFieldConfig, isFormFieldConfig } from "../generator";
 import { camelCaseToHumanReadable } from "../utils/camelCaseToHumanReadable";
 import { findQueryTypeOrThrow } from "../utils/findQueryType";
 import { Imports } from "../utils/generateImportsCode";
 import { isFieldOptional } from "../utils/isFieldOptional";
 import { findFieldByName, GenerateFieldsReturn } from "./generateFields";
+
+function convertGqlScalarToTypescript(scalarName: string) {
+    if (scalarName === "String" || scalarName === "ID") {
+        return "string";
+    } else if (scalarName === "Boolean") {
+        return "boolean";
+    } else if (scalarName === "Int" || scalarName === "Float") {
+        return "number";
+    } else if (scalarName === "DateTime") {
+        return "Date";
+    } else if (scalarName === "JSONObject") {
+        return "unknown";
+    } else {
+        return "unknown";
+    }
+}
+
+function getTypeInfo(arg: IntrospectionInputValue | undefined, gqlIntrospection: IntrospectionQuery) {
+    if (!arg) return undefined;
+
+    let typeKind = undefined;
+    let typeClass = "unknown";
+    let required = false;
+    let type = arg.type;
+    let introspectedType = undefined;
+
+    if (type.kind === "NON_NULL") {
+        required = true;
+        type = type.ofType;
+    }
+    if (type.kind === "INPUT_OBJECT") {
+        typeClass = type.name;
+        typeKind = type.kind;
+        introspectedType = gqlIntrospection.__schema.types.find((type) => type.name === typeClass);
+    } else if (type.kind === "ENUM") {
+        typeClass = type.name;
+        typeKind = type.kind;
+    } else if (type.kind === "SCALAR") {
+        typeClass = type.name;
+        typeKind = type.kind;
+    } else {
+        throw new Error(`Resolving kind ${type.kind} currently not supported.`);
+    }
+    return {
+        required,
+        typeKind,
+        typeClass,
+        introspectedType,
+    };
+}
 
 export function generateFormField({
     gqlIntrospection,
@@ -57,6 +108,8 @@ export function generateFormField({
     const readOnlyPropsWithLock = `${readOnlyProps} ${endAdornmentWithLockIconProp}`;
 
     const imports: Imports = [];
+    const props: Prop[] = [];
+
     const defaultFormValuesConfig: GenerateFieldsReturn["formValuesConfig"][0] = {
         destructFromFormValues: config.virtual ? name : undefined,
     };
@@ -328,10 +381,113 @@ export function generateFormField({
         const rootQuery = config.rootQuery; //TODO we should infer a default value from the gql schema
         const queryName = `${rootQuery[0].toUpperCase() + rootQuery.substring(1)}Select`;
         const rootQueryType = findQueryTypeOrThrow(rootQuery, gqlIntrospection);
-        const rootQueryFilterType = rootQueryType.args.find((arg) => arg.name === "filter");
 
-        const filterField = config.filterField?.name ? findFieldByName(config.filterField.name, formConfig.fields) : undefined;
-        if (filterField) {
+        formFragmentField = `${name} { id ${labelField} }`;
+
+        const filterConfig = config.filter
+            ? (() => {
+                  const filterField = config.filter.type === "field" ? findFieldByName(config.filter.name, formConfig.fields) : undefined;
+                  if (config.filter.type === "field") {
+                      if (!filterField) {
+                          throw new Error(
+                              `Field ${String(config.name)}: No field with name "${
+                                  config.filter?.name
+                              }" referenced as filterField found in form-config.`,
+                          );
+                      }
+                      if (!isFormFieldConfig(filterField)) {
+                          throw new Error(
+                              `Field ${String(config.name)}: Field with name "${config.filter?.name}" referenced as filterField is no FormField.`,
+                          );
+                      }
+                  }
+
+                  const gqlName = config.filter.gqlName ?? config.filter.name;
+
+                  // try to find arg used to filter by checking names of root-props and filter-prop-fields
+                  let gqlVarType = getTypeInfo(
+                      rootQueryType.args.find((arg) => arg.name === gqlName),
+                      gqlIntrospection,
+                  );
+                  let gqlVarName = undefined;
+                  let gqlVarAssignment = undefined;
+                  const filterVar = filterField
+                      ? `values.${filterField.type === "asyncSelect" ? `${String(filterField.name)}?.id` : String(filterField.name)}`
+                      : `${config.filter.name}`;
+                  let filterVarType = "unknown";
+
+                  if (gqlVarType) {
+                      // there is a root-prop with same name, so the dev probably wants to filter with this prop
+                      gqlVarName = gqlName;
+                      gqlVarAssignment = filterVar;
+                      if (gqlVarType.typeKind === "INPUT_OBJECT" || gqlVarType.typeKind === "ENUM") {
+                          filterVarType = `GQL${gqlVarType.typeClass}`;
+                          imports.push({
+                              name: filterVarType,
+                              importPath: "@src/graphql.generated",
+                          });
+                      } else if (gqlVarType.typeKind === "SCALAR") {
+                          filterVarType = convertGqlScalarToTypescript(gqlVarType.typeClass);
+                      }
+                  } else {
+                      // no root-prop with same name, check filter-prop-fields
+                      gqlVarType = getTypeInfo(
+                          rootQueryType.args.find((arg) => arg.name === "filter"),
+                          gqlIntrospection,
+                      );
+                      if (gqlVarType) {
+                          gqlVarName = "filter";
+                          gqlVarAssignment = `{ ${gqlName}: { equal: ${filterVar} } }`;
+
+                          // get type of field.equal in filter-arg used for filtering
+                          if (!gqlVarType.introspectedType || gqlVarType.introspectedType.kind !== "INPUT_OBJECT") {
+                              throw new Error(`Field ${String(config.name)}: Type of filter is no object-type.`);
+                          }
+                          const gqlFilterInputType = getTypeInfo(
+                              gqlVarType.introspectedType.inputFields.find((inputField) => inputField.name === gqlName),
+                              gqlIntrospection,
+                          );
+                          if (!gqlFilterInputType?.introspectedType || gqlFilterInputType.introspectedType.kind !== "INPUT_OBJECT") {
+                              throw new Error(
+                                  `Field ${String(
+                                      config.name,
+                                  )}: Type of filter.${gqlName} is no object-type, but needs to be e.g. StringFilter-type.`,
+                              );
+                          }
+                          const equalFieldType = getTypeInfo(
+                              gqlFilterInputType.introspectedType.inputFields.find((inputField) => inputField.name === "equal"),
+                              gqlIntrospection,
+                          );
+                          if (!equalFieldType) {
+                              throw new Error(
+                                  `Field ${String(config.name)}: Field filter.${gqlName}.equal does not exist but is required for filtering.`,
+                              );
+                          }
+                          filterVarType = convertGqlScalarToTypescript(equalFieldType.typeClass);
+                      } else {
+                          throw new Error(
+                              `Neither filter-prop nor root-prop with name: ${gqlName} for asyncSelect-query not found. Consider setting filterField.gqlVarName explicitly.`,
+                          );
+                      }
+                  }
+
+                  if (config.filter.type === "prop") {
+                      props.push({
+                          name: config.filter.name,
+                          optional: false,
+                          type: filterVarType,
+                      });
+                  }
+
+                  return {
+                      filterField,
+                      gqlVarType,
+                      gqlVarName,
+                      gqlVarAssignment,
+                  };
+              })()
+            : undefined;
+        if (filterConfig) {
             imports.push({ name: "OnChangeField", importPath: "@comet/admin" });
             finalFormConfig = { subscription: { values: true }, renderProps: { values: true, form: true } };
         }
@@ -377,6 +533,7 @@ export function generateFormField({
         }
 
         formValueToGqlInputCode = !config.virtual ? `${name}: formValues.${name}?.id,` : ``;
+
         imports.push({
             name: `GQL${queryName}Query`,
             importPath: `./${baseOutputFilename}.generated`,
@@ -395,39 +552,31 @@ export function generateFormField({
                 loadOptions={async () => {
                     const { data } = await client.query<GQL${queryName}Query, GQL${queryName}QueryVariables>({
                         query: gql\`query ${queryName}${
-            filterField && rootQueryFilterType && rootQueryFilterType.type.kind === "INPUT_OBJECT"
-                ? `($filter: ${rootQueryFilterType.type.name})`
-                : ``
+            filterConfig ? `($${filterConfig.gqlVarName}: ${filterConfig.gqlVarType.typeClass}${filterConfig.gqlVarType.required ? `!` : ``})` : ``
         } {
-                            ${rootQuery}${filterField && rootQueryFilterType ? `(filter: $filter)` : ``} {
+                            ${rootQuery}${filterConfig ? `(${filterConfig.gqlVarName}: $${filterConfig.gqlVarName})` : ``} {
                                 nodes {
                                     id
                                     ${labelField}
                                 }
                             }
                         }\`${
-                            filterField && rootQueryFilterType && config.filterField
-                                ? `, variables: { ${(config.filterField?.gqlVarType === "filter"
-                                      ? `filter: { gqlVarName: { equal: valuesVar } }`
-                                      : `gqlVarName: valuesVar`
-                                  )
-                                      .replace(`gqlVarName`, config.filterField.gqlVarName)
-                                      .replace(
-                                          `valuesVar`,
-                                          `values.${
-                                              filterField.type === "asyncSelect" ? `${String(filterField.name)}?.id` : String(filterField.name)
-                                          }`,
-                                      )} }`
+                            filterConfig
+                                ? `, variables: { ${
+                                      filterConfig.gqlVarName == filterConfig.gqlVarAssignment
+                                          ? filterConfig.gqlVarName
+                                          : `${filterConfig.gqlVarName}: ${filterConfig.gqlVarAssignment}`
+                                  } }`
                                 : ``
                         }
                     });
                     return data.${rootQuery}.nodes;
                 }}
                 getOptionLabel={(option) => option.${labelField}}
-                ${filterField ? `disabled={!values?.${String(filterField.name)}}` : ``}
+                ${filterConfig && filterConfig.filterField ? `disabled={!values?.${String(filterConfig.filterField.name)}}` : ``}
             />${
-                filterField
-                    ? `<OnChangeField name="${String(filterField.name)}">
+                filterConfig && filterConfig.filterField
+                    ? `<OnChangeField name="${String(filterConfig.filterField.name)}">
                             {(value, previousValue) => {
                                 if (value.id !== previousValue.id) {
                                     form.change("${String(config.name)}", undefined);
@@ -439,9 +588,15 @@ export function generateFormField({
     } else {
         throw new Error(`Unsupported type`);
     }
+
+    if (config.optionalRender) {
+        code = `{ show${name[0].toUpperCase() + name.substring(1)} && ${code} }`;
+        props.push({ name: `show${name[0].toUpperCase() + name.substring(1)}`, type: `boolean`, optional: true });
+    }
+
     return {
-        code: config.optionalRender ? `{ show${name[0].toUpperCase() + name.substring(1)} && ${code} }` : code,
-        props: config.optionalRender ? [{ name: `show${name[0].toUpperCase() + name.substring(1)}`, type: `boolean`, optional: true }] : [],
+        code,
+        props,
         hooksCode,
         formValueToGqlInputCode,
         formFragmentFields: [formFragmentField],