+++ packages/admin/admin-rte/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @comet/admin-rte
 
+## 7.2.1
+
+### Patch Changes
+
+-   @comet/admin@7.2.1
+-   @comet/admin-icons@7.2.1
+
 ## 7.2.0
 
 ### Patch Changes